+++ karma.conf.js
@@ -1,71 +1,40 @@
-var webpack = require('webpack');
+const webpackConfig = require('./webpack.config.js');
 
-module.exports = function(config) {
+module.exports = function karma(config) {
   config.set({
-
-    // base path that will be used to resolve all patterns (eg. files, exclude)
-    basePath: '',
-
-
-    // frameworks to use
-    // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
     frameworks: ['mocha', 'chai', 'sinon'],
+    reporters: ['mocha'],
 
-
-    // list of files / patterns to load in the browser
-    files: [
-      'spec/**/*.spec.js'
-    ],
-
-    // list of files to exclude
-    exclude: [
-    ],
-
-    // preprocess matching files before serving them to the browser
-    // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {
-      'spec/**/*.spec.js': ['webpack']
-    },
-
-    webpack: require('./webpack.config.js'),
-    webpackMiddleware: {
-      stats: 'errors-only'
-    },
-
-    // test results reporter to use
-    // possible values: 'dots', 'progress'
-    // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-    reporters: ['progress'],
-
-
-    // web server port
-    port: 9876,
-
-
-    // enable / disable colors in the output (reporters and logs)
+    port: 52987, // karma web server port
     colors: true,
-
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     logLevel: config.LOG_INFO,
-
-
-    // enable / disable watching file and executing tests whenever any file changes
+    browsers: ['ChromeHeadless_NoSandbox'],
     autoWatch: false,
+    singleRun: true, // Karma captures browsers, runs the tests and exits
+    concurrency: Infinity,
 
+    files: ['spec/**/*.spec.js'],
 
-    // start these browsers
-    // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    browsers: ['PhantomJS'],
+    preprocessors: {
+      './spec/**/*.spec.js': ['webpack']
+    },
 
+    customLaunchers: {
+      ChromeHeadless_NoSandbox: {
+        base: 'ChromeHeadless',
+        flags: ['--no-sandbox'],
+        displayName: 'Chrome Headless w/o Sandbox'
+      }
+    },
 
-    // Continuous Integration mode
-    // if true, Karma captures browsers, runs the tests and exits
-    singleRun: true,
+    webpack: webpackConfig,
 
-    // Concurrency level
-    // how many browser should be started simultaneous
-    concurrency: Infinity
-  })
-}
+    webpackMiddleware: {
+      noInfo: true
+    },
+
+    mochaReporter: {
+      showDiff: true
+    }
+  });
+};