+++ src/user.js
@@ -1,11 +1,42 @@
 /**
- * User stores fields related to the current extension user, either a viewer or the broadcaster.
+ * Stores fields related to the current extension user, either a viewer or the broadcaster.
  * These fields are automatically updated by the SDK.
  */
 export default class User {
   /**
+   * Defines the current user's role on Twitch relative to the current channel being
+   * viewed. May be "viewer" if the user is simply viewing the channel, "moderator"
+   * if the user is a moderator of the channel or "broadcaster" if the user is also
+   * the broadcaster of the channel.
+   *
+   * @since 1.0.3
+   */
+  static get Roles() {
+    return {
+      Viewer: 'viewer',
+      Broadcaster: 'broadcaster',
+      Moderator: 'moderator'
+    };
+  }
+
+  /**
+   * Defines the video mode for the current user. This may be "default" for the default
+   * windowed viewing experience on Twitch, "fullscreen" for the fullscreen, video-only
+   * mode or "theatre" for the video full window-width.
+   *
+   * @since 1.0.3
+   */
+  static get VideoModes() {
+    return {
+      Default: 'default',
+      Fullscreen: 'fullscreen',
+      Theatre: 'theatre'
+    };
+  }
+
+  /**
    * @since 1.0.0
-   * @param {object} auth - An auth token usable by this user for backend requests.
+   * @param {Object} auth - An auth token usable by this user for backend requests.
    */
   constructor(auth) {
     /**
@@ -20,7 +51,7 @@ export default class User {
      * twitchJWT holds the raw JWT response from the Twitch Extension SDK.
      *
      * @since 1.0.0
-     * @type {object}
+     * @type {Object}
      */
     this.twitchJWT = auth.token;
 
@@ -72,12 +103,12 @@ export default class User {
     this.visualizationID = '';
 
     /**
-     * role is the current user's role in the extension. May be one of ['viewer', 'config'].
+     * role is the current user's role in the extension. May be one of {@link Roles}.
      *
      * @since 1.0.0
      * @type {string}
      */
-    this.role = 'viewer';
+    this.role = User.Roles.Viewer;
 
     /**
      * ip is the current user's IP address. May be an empty string if undetectable.
@@ -96,12 +127,12 @@ export default class User {
     this.game = '';
 
     /**
-     * Current video mode: One of ['default', fullscreen', 'theatre'].
+     * User's current video mode. One of {@link VideoModes}.
      *
      * @since 1.0.0
      * @type {string}
      */
-    this.videoMode = 'default';
+    this.videoMode = User.VideoModes.Default;
 
     /**
      * Current video bitrate. Null if no video or unknown.
@@ -120,7 +151,7 @@ export default class User {
     this.latency = null;
 
     /**
-     * Current buffer size of the viewer's player. Null if no video or uknown.
+     * Current buffer size of the viewer's player. Null if no video or unknown.
      *
      * @since 1.0.0
      * @type {null|number}
@@ -136,7 +167,7 @@ export default class User {
    * Attempts to parse the provided JWT and persist any found information in store.
    * @since 1.0.0
    *
-   * @param {object} jwt - The auth JWT token as returned from the auth harness.
+   * @param {Object} jwt - The auth JWT token as returned from the auth harness.
    */
   extractJWTInfo(jwt) {
     try {
@@ -171,7 +202,7 @@ export default class User {
    * Stores values from a new auth token in the local store.
    * @since 1.0.0
    *
-   * @param {object} auth - An auth JWT with updated user information.
+   * @param {Object} auth - An auth JWT with updated user information.
    */
   updateAuth(auth) {
     this.twitchJWT = auth.token;