+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 
 ## Change Log
 
+## 2017-09-21
+v1.0.3
+
+### Updated
+- Additional documentation.
+- Fixed UMD ES6 class fighting.
+
+## 2017-09-11
+v1.0.2
+
+### Updated
+- Moved xhr-promise lib into the repo to work around a CSP enforcement issue on Twitch.
+
+
 ## 2017-09-08
+v1.0.1
 
 ### Updated
 - Removed analytics access to localStorage (unused and unneeded).
@@ -14,5 +29,6 @@
 
 
 ## 2017-08-31
+v1.0.0
 
-Initial release
+- Initial release